small fixes in 1_1, move pains.py content in unwanted_substructures.py
